refactor(settings): split rule authoring action catalog out of ruleauthoring.cpp
