+++ tests/unit/CMakeLists.txt
@@ -1142,6 +1142,7 @@ add_executable(test_rule_model
     # rulemodel's action summary now shares its boolean polarity phrasing with
     # the editor via RuleAuthoring::boolActionStateLabel.
     ${CMAKE_SOURCE_DIR}/src/settings/rules/ruleauthoring.cpp
+    ${CMAKE_SOURCE_DIR}/src/settings/rules/ruleauthoring_actions.cpp
 )
 target_link_libraries(test_rule_model
     PRIVATE
@@ -1165,6 +1166,7 @@ add_executable(test_rule_controller
     ${CMAKE_SOURCE_DIR}/src/settings/rules/rulecontroller_lookups.cpp
     ${CMAKE_SOURCE_DIR}/src/settings/rules/rulecontroller_baseline.cpp
     ${CMAKE_SOURCE_DIR}/src/settings/rules/ruleauthoring.cpp
+    ${CMAKE_SOURCE_DIR}/src/settings/rules/ruleauthoring_actions.cpp
     ${CMAKE_SOURCE_DIR}/src/settings/rules/ruletemplates.cpp
 )
 target_link_libraries(test_rule_controller
